refactor: move services/mount/ to bricks/mount/ (LEGO architecture)
Move mount module from legacy services/ bucket to bricks/ tier per
KERNEL-ARCHITECTURE.md. Fix brick boundary violations by replacing
forbidden TYPE_CHECKING imports with Any and using importlib for
check_permission runtime import. Update all consumer imports across
factory, tests, and protocol docstrings.
