+++ src/nexus/bricks/mount/__init__.py
@@ -1,0 +1,14 @@
+"""Mount brick — BRICK tier.
+
+Canonical location for mount lifecycle services.
+"""
+
+from nexus.bricks.mount.mount_core_service import MountCoreService
+from nexus.bricks.mount.mount_manager import MountManager
+from nexus.bricks.mount.mount_service import MountService
+
+__all__ = [
+    "MountCoreService",
+    "MountManager",
+    "MountService",
+]
