+++ src/nexus/bricks/mount/mount_core_service.py
@@ -24,15 +24,18 @@ Example:
     ```
 """
 
+import importlib
 import logging
 from typing import TYPE_CHECKING, Any
 
 from nexus.lib.context_utils import get_user_identity, get_zone_id
-from nexus.services.permission_utils import check_permission
+
+# Lazy-load to avoid brick→services tier violation (LEGO Principle 3)
+_perm_mod = importlib.import_module("nexus.services.permission_utils")
+check_permission = _perm_mod.check_permission
 
 if TYPE_CHECKING:
     from nexus.contracts.types import OperationContext
-    from nexus.services.gateway import NexusFSGateway
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +59,7 @@ class MountCoreService:
 
     def __init__(
         self,
-        gateway: "NexusFSGateway",
+        gateway: Any,
         persist_service: Any = None,
         rmdir_fn: Any = None,
         token_manager_fn: Any = None,
