+++ src/nexus/services/__init__.py
@@ -24,7 +24,6 @@ Phase 2: Core Refactoring (Issue #988)
 import importlib as _il
 
 from nexus.contracts.types import ContextIdentity, extract_context_identity
-from nexus.services.mount.mount_service import MountService
 from nexus.services.oauth.oauth_service import OAuthService
 from nexus.services.scheduler import SchedulerService
 from nexus.services.search.search_service import SearchService
@@ -34,6 +33,7 @@ from nexus.services.versioning.version_service import VersionService
 # Brick re-exports via importlib to avoid services→bricks tier violation (import-linter)
 LLMService = _il.import_module("nexus.bricks.llm.llm_service").LLMService
 MCPService = _il.import_module("nexus.bricks.mcp.mcp_service").MCPService
+MountService = _il.import_module("nexus.bricks.mount.mount_service").MountService
 ReBACService = _il.import_module("nexus.bricks.rebac.rebac_service").ReBACService
 __all__ = [
     "SearchService",
