--- src/nexus/services/mount/__init__.py
@@ -1,14 +1,0 @@
-"""Mount service domain -- BRICK tier.
-
-Canonical location for mount lifecycle services.
-"""
-
-from nexus.services.mount.mount_core_service import MountCoreService
-from nexus.services.mount.mount_manager import MountManager
-from nexus.services.mount.mount_service import MountService
-
-__all__ = [
-    "MountCoreService",
-    "MountManager",
-    "MountService",
-]
